browser/components/extensions/test/mochitest/test_ext_history_redirect.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for simple WebExtension</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SpawnTask.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/ExtensionTestUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>

<script type="text/javascript">
"use strict";

const REDIRECT_URL = new URL("redirection.sjs", location).href;

function waitForLoad(win) {
  return new Promise(resolve => {
    win.addEventListener("load", function() {
      resolve();
    }, {capture: true, once: true});
  });
}

add_task(async function history_redirect() {
  function background() {
    browser.test.onMessage.addListener(async (msg, url) => {
      switch (msg) {
        case "delete-all": {
          let results = await browser.history.deleteAll();
          browser.test.sendMessage("delete-all-result", results);
          break;
        }
        case "search": {
          let results = await browser.history.search({text: url, startTime: new Date(0)});
          browser.test.sendMessage("search-result", results);
          break;
        }
        case "get-visits": {
          let results = await browser.history.getVisits({url});
          browser.test.sendMessage("get-visits-result", results);
          break;
        }
      }
    });

    browser.test.sendMessage("ready");
  }

  let extensionData = {
    manifest: {
      permissions: [
        "history",
      ],
    },
    background,
  };

  let extension = ExtensionTestUtils.loadExtension(extensionData);

  await Promise.all([extension.startup(), extension.awaitMessage("ready")]);
  info("extension loaded");

  extension.sendMessage("delete-all");
  await extension.awaitMessage("delete-all-result");

  let win = window.open(REDIRECT_URL);
  await waitForLoad(win);

  extension.sendMessage("search", REDIRECT_URL);
  let results = await extension.awaitMessage("search-result");
  is(results.length, 1, "search returned expected length of results");

  extension.sendMessage("get-visits", REDIRECT_URL);
  let visits = await extension.awaitMessage("get-visits-result");
  is(visits.length, 1, "getVisits returned expected length of visits");

  // cleanup phase
  win.close();

  await extension.unload();
  info("extension unloaded");
});
</script>

</body>
</html>
